Working commit! This commit works now
Although there are still multiple TODO comments, and some unexplained
behavior, this is the first commit that uses a LFSR as a Program Counter
and works, it books up an eForth image. The problem appeared to be
caused by not allocating enough space after that VM, why this is needs
looking at.

Some more TODO comments have been added which should aid in simplifying
the hardware (although not the software) implementation if one is made.