Changed PC size, fixed startup code
There are some bugs that you think "how the hell did that ever work?",
this commit fixes one of those, the start up code used the completely
wrong instruction after changing the instruction ordering and just
happened to work anyway (changing BXXX to 8XXX).

The Program Counter size has been reduced to 8 bits, as that is all that
is needed for the C virtual machine.